chore: restructure App UI components into feature-folder hierarchy
Closes #129

Move all Razor components from a flat Components/ layout
(Pages/, Layout/, Dialogs/, Shared/) into a feature-folder
hierarchy that mirrors the back-end layer restructure:

  Shell/           - App entry, routing, main layout
  Features/        - Feature-scoped pages, components, dialogs
    About/Pages/
    Audit/Pages/
    BoardRules/Pages/
    Common/Pages/  - Error and NotFound pages
    Dashboard/Pages/
    Labels/{Pages,Components,Dialogs}/
    Migration/{Pages,Components}/
    Repositories/Pages/
    Triage/Pages/
    Workflows/Pages/
  Shared/Components/ - Cross-feature shared components

Also update:
- All C# namespaces in code-behind and model files
- _Imports.razor global using directives
- Routes.razor type references (global:: prefix)
- Program.cs shell namespace reference
- ReconnectModal.razor static asset path
- 6 bUnit test files with new namespaces
- plan/BACKLOG.md to record completion

Build: 0 errors, 0 warnings
Tests: 175 passed, 0 failed
