+++ plan/BACKLOG.md
@@ -234,6 +234,7 @@ Labels: `type/chore`, `area/infrastructure`
 - [x] Set up Bicep infrastructure for Azure App Service, Key Vault, and managed identity. _(done — Bicep baseline complete with managed identity RBAC, configurable Key Vault secret name, purge protection, health-check path, and optional CIDR inbound access restrictions (SKU-gated; unsupported on F1); guided PowerShell deployment script `infra/Deploy-SoloDevBoardInfra.ps1` added with auto-detected caller IP defaulting, F1 compatibility guard, and post-deployment next-step guidance; see issue #104.)_
 - [x] Configure OIDC authentication for GitHub Actions to Azure (no long-lived credentials). _(done — `cd.yml` already uses OIDC with least-privilege permissions; OIDC trust prerequisites and protected-environment documentation added to `infra/README.md`; see issue #105.)_
 - [x] Chore: Wire GitHub Actions CD pipeline end-to-end with Azure OIDC deployment identity (user-assigned managed identity, federated credential, Contributor role at resource-group scope), surface live deployment URL in GitHub Environments UI, and refactor `Deploy-SoloDevBoardInfra.ps1` with opt-in IP hardening, git-remote repo auto-detection, resource name suffix support for self-hosters, and opt-in verbose next-steps guidance. _(No issue — done, 2026-03-17; branch `get-deploying`.)_
+- [x] Chore: App UI components restructured from a flat `Pages/`, `Layout/`, `Dialogs/`, `Shared/` layout into a feature-folder hierarchy (`Shell/`, `Features/{FeatureName}/{Pages,Components,Dialogs}/`, `Shared/Components/`) with aligned namespaces, consistent with the back-end layer restructure, so that the App project is easier to navigate and scales for future feature delivery. _(Issue #129 — done, 2026-03-17.)_
 - [ ] Implement response caching for GitHub API calls to respect rate limits. _(#108)_
 - [ ] Add health check endpoints for Azure App Service monitoring. _(#106)_
 - [ ] Configure logging with structured output (Azure Application Insights integration). _(#107)_
