+++ plan/wireframes/README.md
@@ -1,0 +1,13 @@
+# Wireframe Reference Documents
+
+This directory contains planning-only wireframe references for key SoloDevBoard pages. These artefacts are intended to guide implementation discussions, clarify user goals, and document layout and interaction patterns. No source code is included.
+
+## How to Use
+- Review each wireframe before proposing UI changes or new features.
+- Use these documents to align on user goals, accessibility, and responsive behaviour.
+- Reference the ASCII wireframes and interaction notes during implementation planning.
+
+## Wireframes
+- [repositories-wireframe.md](repositories-wireframe.md): Repositories page wireframe and interaction notes.
+- [label-manager-wireframe.md](label-manager-wireframe.md): Label Manager page wireframe, mode separation rationale, and tabbed IA.
+- [audit-dashboard-wireframe.md](audit-dashboard-wireframe.md): Audit Dashboard wireframe, KPI cards, health indicators, and filter surface.
