+++ src/App/SoloDevBoard.App/Components/Features/BoardRules/Pages/BoardRules.razor
