+++ src/App/SoloDevBoard.App/Components/Features/Common/Pages/NotFound.razor
@@ -1,5 +1,5 @@
 ﻿@page "/not-found"
-@layout Layout.MainLayout
+@layout MainLayout
 
 <PageTitle>Not Found</PageTitle>
