+++ src/App/SoloDevBoard.App/Components/Features/Labels/Components/LabelPreviewTable.razor.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Components;
 using SoloDevBoard.Application.Services.Labels;
 
-namespace SoloDevBoard.App.Components.Pages;
+namespace SoloDevBoard.App.Components.Features.Labels.Components;
 
 /// <summary>
 /// Code-behind for the LabelPreviewTable component.
