--- src/App/SoloDevBoard.App/Components/Routes.razor
@@ -1,6 +1,0 @@
-﻿<Router AppAssembly="typeof(Program).Assembly" NotFoundPage="typeof(Pages.NotFound)">
-    <Found Context="routeData">
-        <RouteView RouteData="routeData" DefaultLayout="typeof(Layout.MainLayout)" />
-        <FocusOnNavigate RouteData="routeData" Selector="h1, .mud-typography-h4, .mud-typography-h5" />
-    </Found>
-</Router>
