+++ src/App/SoloDevBoard.App/Components/Shell/App.razor
