+++ src/App/SoloDevBoard.App/Components/Shell/Layout/ReconnectModal.razor
@@ -1,4 +1,4 @@
-﻿<script type="module" src="@Assets["Components/Layout/ReconnectModal.razor.js"]"></script>
+﻿<script type="module" src="@Assets["Components/Shell/Layout/ReconnectModal.razor.js"]"></script>
 
 <dialog id="components-reconnect-modal" data-nosnippet>
     <div class="components-reconnect-backdrop">
