+++ src/App/SoloDevBoard.App/Components/Shell/Routes.razor
@@ -1,0 +1,6 @@
+﻿<Router AppAssembly="typeof(Program).Assembly" NotFoundPage="typeof(global::SoloDevBoard.App.Components.Features.Common.Pages.NotFound)">
+    <Found Context="routeData">
+        <RouteView RouteData="routeData" DefaultLayout="typeof(global::SoloDevBoard.App.Components.Shell.Layout.MainLayout)" />
+        <FocusOnNavigate RouteData="routeData" Selector="h1, .mud-typography-h4, .mud-typography-h5" />
+    </Found>
+</Router>
