+++ src/App/SoloDevBoard.App/Components/_Imports.razor
@@ -11,8 +11,11 @@
 @using Microsoft.JSInterop
 @using SoloDevBoard.App
 @using SoloDevBoard.App.Components
-@using SoloDevBoard.App.Components.Shared
-@using SoloDevBoard.App.Components.Layout
+@using SoloDevBoard.App.Components.Shell.Layout
+@using SoloDevBoard.App.Components.Shared.Components
+@using SoloDevBoard.App.Components.Features.Labels.Components
+@using SoloDevBoard.App.Components.Features.Labels.Dialogs
+@using SoloDevBoard.App.Components.Features.Migration.Components
 @using SoloDevBoard.Application.Services.Audit
 @using SoloDevBoard.Application.Services.Common
 @using SoloDevBoard.Application.Services.Labels
