+++ tests/App.Tests/SoloDevBoard.App.Tests/MigrationTests.cs
@@ -3,8 +3,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using MudBlazor;
 using MudBlazor.Services;
-using SoloDevBoard.App.Components.Pages;
-using SoloDevBoard.App.Components.Shared;
+using SoloDevBoard.App.Components.Features.Migration.Pages;
+using SoloDevBoard.App.Components.Shared.Components;
 using SoloDevBoard.Application.Services.Labels;
 using SoloDevBoard.Application.Services.Migration;
 using SoloDevBoard.Application.Services.Repositories;
